third_party/WebKit/LayoutTests/typedcssom/cssSkew-expected.txt

This is a testharness.js-based test.
PASS (ax, ay) values are correct for skew(0deg, 0deg) 
PASS (ax, ay) values are correct for skew(1deg, 2deg) 
PASS (ax, ay) values are correct for skew(-2deg, -4deg) 
PASS (ax, ay) values are correct for skew(3.4deg, 2.7deg) 
PASS (ax, ay) values are correct for skew(1rad, 0deg) 
PASS (ax, ay) values are correct for skew(0deg, 1rad) 
PASS is2D value is correct for skew(0deg, 0deg) 
PASS is2D value is correct for skew(1deg, 2deg) 
PASS is2D value is correct for skew(-2deg, -4deg) 
PASS is2D value is correct for skew(3.4deg, 2.7deg) 
PASS is2D value is correct for skew(1rad, 0deg) 
PASS is2D value is correct for skew(0deg, 1rad) 
FAIL toString is correct for skew(0deg, 0deg) assert_equals: expected "skew(0deg, 0deg)" but got ""
FAIL toString is correct for skew(1deg, 2deg) assert_equals: expected "skew(1deg, 2deg)" but got ""
FAIL toString is correct for skew(-2deg, -4deg) assert_equals: expected "skew(-2deg, -4deg)" but got ""
FAIL toString is correct for skew(3.4deg, 2.7deg) assert_equals: expected "skew(3.4deg, 2.7deg)" but got ""
FAIL toString is correct for skew(1rad, 0deg) assert_equals: expected "skew(1rad, 0deg)" but got ""
FAIL toString is correct for skew(0deg, 1rad) assert_equals: expected "skew(0deg, 1rad)" but got ""
PASS Invalid arguments to constructor throws an exception. 
PASS asMatrix is constructed correctly for skew(0deg, 0deg) 
FAIL asMatrix is constructed correctly for skew(1deg, 2deg) assert_approx_equals: expected 0.03492076949174773 +/- 0.000001 but got 0
FAIL asMatrix is constructed correctly for skew(-2deg, -4deg) assert_approx_equals: expected -0.06992681194351041 +/- 0.000001 but got 0
FAIL asMatrix is constructed correctly for skew(3.4deg, 2.7deg) assert_approx_equals: expected 0.04715880287748047 +/- 0.000001 but got 0
FAIL asMatrix is constructed correctly for skew(1rad, 0deg) assert_approx_equals: expected 1.5574077246549023 +/- 0.000001 but got 0
FAIL asMatrix is constructed correctly for skew(0deg, 1rad) assert_approx_equals: expected 1.5574077246549023 +/- 0.000001 but got 0
PASS Setting ax and ay with valid CSSUnitValues 
PASS Setting ax with invalid values 
PASS Setting ay with invalid values 
Harness: the test ran to completion.
